Killing my darlings
I learned from this post and I think it makes a valid point. But I have
issues with it, too, that are strong enough to warrant removing it from
the list.

Ideally this list would be essays I think about, or have thought about,
and not essays that define my thinking, with which I agree on every
point made. Unfortunately, I think that distinction gets lost in the
HTML on the page.